Extract WPInstall class
